+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Fixed handling of errors in recursive QCJSON creator operations.
 - Support for xtb optimizations, single-point energies, and molecular dynamics. (xtb)
 - Support for additional functionals. (ORCA)
+- Support for coupled cluster calculations. (ORCA)
 
 ## [0.2.0] 2021-02-09